requirements.txt
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
Pygments==2.11.2
python-dotenv==0.19.2